feat: 주문 로직에 쿠폰 적용 기능 추가

+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderInfo.java
@@ -31,7 +31,7 @@ public record OrderInfo(
         return new OrderInfo(
                 order.getId(),
                 order.getUser().getUserId(),
-                order.getTotalPrice(),
+                order.getTotalPrice().intValue(),
                 order.getStatus(),
                 itemInfos,
                 externalSendInfo
